move swap varant tests
