+++ contracts/transmuter/src/swap/mod.rs
@@ -20,9 +20,7 @@ mod tests {
     use crate::{
         asset::Asset,
         corruptable::Corruptable,
-        swap::common::{
-            Entrypoint, SwapExactAmountInResponseData, SwapExactAmountOutResponseData, SwapVariant,
-        },
+        swap::common::{Entrypoint, SwapExactAmountInResponseData, SwapExactAmountOutResponseData},
         transmuter_pool::AssetGroup,
         ContractError,
     };
@@ -41,32 +39,6 @@ mod tests {
     use osmosis_test_tube::cosmrs::proto::prost::Message;
     use rstest::rstest;
     use transmuter_math::rebalancing::config::RebalancingConfig;
-
-    #[rstest]
-    #[case("denom1", "denom2", Ok(SwapVariant::TokenToToken))]
-    #[case("denom2", "denom1", Ok(SwapVariant::TokenToToken))]
-    #[case("denom1", "denom1", Err(ContractError::SameDenomNotAllowed {
-        denom: "denom1".to_string()
-    }))]
-    #[case("denom1", "alloyed", Ok(SwapVariant::TokenToAlloyed))]
-    #[case("alloyed", "denom1", Ok(SwapVariant::AlloyedToToken))]
-    #[case("alloyed", "alloyed", Err(ContractError::SameDenomNotAllowed {
-        denom: "alloyed".to_string()
-    }))]
-    fn test_swap_variant(
-        #[case] denom1: &str,
-        #[case] denom2: &str,
-        #[case] res: Result<SwapVariant, ContractError>,
-    ) {
-        let mut deps = cosmwasm_std::testing::mock_dependencies();
-        let transmuter = Transmuter::new();
-        transmuter
-            .alloyed_asset
-            .set_alloyed_denom(&mut deps.storage, &"alloyed".to_string())
-            .unwrap();
-
-        assert_eq!(transmuter.swap_variant(denom1, denom2, deps.as_ref()), res);
-    }
 
     #[rstest]
     #[case(
